worker/requirements.txt
fastapi
uvicorn[standard]
grpcio
grpcio-tools
aioredis
httpx
prometheus-client
protobuf
opentelemetry-api
opentelemetry-sdk
